+++ lib/aliyun_oss/bucket/acl.ex
@@ -23,16 +23,17 @@ defmodule Aliyun.Oss.Bucket.ACL do
               "Owner" => %{"DislayName" => "11111111", "ID" => "11111111"}
             }
           },
-          headers: [
-            {"Date", "Wed, 05 Dec 2018 02:34:57 GMT"}
-          ]
+          headers: %{
+            "connection" => ["keep-alive],
+            ...
+          }
         }
       }
 
   """
   @spec get(Config.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
   def get(config, bucket) do
-    get_bucket(config, bucket, %{"acl" => nil})
+    get_bucket(config, bucket, query_params: %{"acl" => nil})
   end
 
   @doc """
@@ -40,29 +41,31 @@ defmodule Aliyun.Oss.Bucket.ACL do
 
   ## Examples
 
-      iex> Aliyun.Oss.Bucket.Acl.put("some-bucket", "public-read")
+      iex> Aliyun.Oss.Bucket.ACL.put("some-bucket", "public-read")
       {:ok,
       %Aliyun.Oss.Client.Response{
         data: "",
-        headers: [
-          {"Server", "AliyunOSS"},
-          {"Date", "Fri, 11 Jan 2019 04:43:42 GMT"},
-          {"Content-Length", "0"},
-          {"Connection", "keep-alive"},
-          {"x-oss-request-id", "5C0000000000000000000000"},
-          {"Location", "/some-bucket"},
-          {"x-oss-server-time", "333"}
-        ]
+        headers: %{
+          "connection" => ["keep-alive"],
+          "content-length" => ["0"],
+          "date" => ["Tue, 08 Jul 2025 02:11:14 GMT"],
+          "location" => ["/some-bucket"],
+          "server" => ["AliyunOSS"],
+          "x-oss-request-id" => ["686C00000000000000000D6B"],
+          "x-oss-server-time" => ["479"]
+        }
       }}
-      iex> Aliyun.Oss.Bucket.Acl.put("some-bucket", "invalid-permission")
+      iex> Aliyun.Oss.Bucket.ACL.put("some-bucket", "invalid-permission")
       {:error,
       %Aliyun.Oss.Client.Error{
         parsed_details: %{
           "ArgumentName" => "x-oss-acl",
-          "ArgumentValue" => "invalid-read",
+          "ArgumentValue" => "invalid-permission",
           "Code" => "InvalidArgument",
+          "EC" => "0015-00000204",
           "HostId" => "some-bucket.oss-cn-shenzhen.aliyuncs.com",
           "Message" => "no such bucket access control exists",
+          "RecommendDoc" => "https://api.aliyun.com/troubleshoot?q=0015-00000204,
           "RequestId" => "5C3000000000000000000000"
         },
         body: "<?xml version=\"1.0\" encoding=\"UTF-8\"?>...</xml>",
@@ -73,6 +76,6 @@ defmodule Aliyun.Oss.Bucket.ACL do
   @spec put(Config.t(), String.t(), String.t()) ::
           {:error, error()} | {:ok, Aliyun.Oss.Client.Response.t()}
   def put(config, bucket, acl) do
-    put_bucket(config, bucket, %{"x-oss-acl" => acl, "acl" => acl}, "")
+    put_bucket(config, bucket, "", headers: %{"x-oss-acl" => acl, "acl" => acl})
   end
 end